context-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-engine"
version = "0.1.0"
description = "Microservice for document parsing and chunking"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "minio>=7.2.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "chardet>=5.2.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["context_engine*"]
